requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
google-api-core==0.1.1
google-auth==1.2.0
google-cloud-bigquery==0.28.0
google-cloud-core==0.28.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
websocket-client==0.44.0
